Add return type restriction to `unbuffered_read` and `unbuffered_write`

+++ src/file/preader.cr
@@ -14,7 +14,7 @@ class File::PReader < IO
     @pos = 0
   end
 
-  def unbuffered_read(slice : Bytes) : Int64
+  def unbuffered_read(slice : Bytes) : Int32
     check_open
 
     count = slice.size
